Restructure to classes.
