+++ src/main.kv
@@ -1,3 +1,9 @@
+#:import DistGauge widgets.DistGauge
+#:import AltGauge widgets.AltGauge
+#:import HGauge widgets.HGauge
+#:import VGauge widgets.VGauge
+#:import HeadingGauge widgets.HeadingGauge
+
 <TopNavBar@MDTopAppBar>
     type: "small"
